[refactor] RsData통합 - RsData의 과도한 책임 분리로 인한 Controller 레이어에서 사용하기 불편한 이슈 방지
