+++ src/main/java/org/com/dungeontalk/domain/member/controller/MemberController.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.com.dungeontalk.domain.member.dto.request.RegisterRequest;
 import org.com.dungeontalk.domain.member.entity.Member;
+import org.com.dungeontalk.domain.member.service.MemberService;
 import org.com.dungeontalk.global.rsData.RsData;
-import org.com.dungeontalk.global.rsData.RsDataFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -30,7 +30,7 @@ public class MemberController {
 
         return ResponseEntity.
                 status(HttpStatus.CREATED).
-                body(RsDataFactory.of("회원가입이 정상적으로 완료되었습니다", member));
+                body(RsData.of("회원가입이 정상적으로 완료되었습니다", member));
     }
 
 }
