+++ src/main/java/org/com/dungeontalk/global/exception/GlobalExceptionHandler.java
@@ -3,7 +3,6 @@ package org.com.dungeontalk.global.exception;
 import org.com.dungeontalk.global.exception.customException.MemberException;
 import org.com.dungeontalk.global.rsData.Empty;
 import org.com.dungeontalk.global.rsData.RsData;
-import org.com.dungeontalk.global.rsData.RsDataFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -17,7 +16,7 @@ public class GlobalExceptionHandler {
         ErrorCode errorCode = ex.getErrorCode();
         return ResponseEntity
                 .status(errorCode.getHttpStatus())
-                .body(RsDataFactory.of(errorCode.getErrorCode(), errorCode.getMessage(), new Empty()));
+                .body(RsData.of(errorCode.getErrorCode(), errorCode.getMessage(), new Empty()));
     }
 
 }
